Refactor db and business logic
Move where the logic for grading the questions is into business logic so
following good coding principles.
